move types to '@frontity/file-settings/types'
