+++ packages/file-settings/src/index.ts
@@ -1,3 +1,2 @@
 export { default as getSettings } from "./getSettings";
 export { default as getAllSites } from "./getAllSites";
-export * from "./types";
