+++ packages/file-settings/src/validateSettings.ts
@@ -1,7 +1,7 @@
 import uniq from "ramda/src/uniq";
 import uniqBy from "ramda/src/uniqBy";
 import { Settings } from "@frontity/types";
-import { Package } from "./types";
+import { Package } from "../types";
 
 // This function validates the packages.
 const validatePackages = (packages: (string | Package)[]): void => {
